feat(dns): add Azure Private DNS provider
With this feature, azacme would be able to handle private certificate
issuing and so could be integrated to a private PKI.
To achieve this, the access permissions to the private zones along side
the Key vault have to be set to azacme.
The server option is set to the private ACME server which also have the
permissions to access to the private zones.
The dns provider option is set to AzurePrivate and the dns zone is
set to a private one.
When all these conditions are met, the user is able to order or renew
private certifcates very easily.